requirements.txt
click==8.0.3
tweepy==4.5.0
pyYAML==5.3